Add instructions for how to debug tests.
Use `--inspect` to get a debugger working.

https://nodejs.org/en/docs/guides/debugging-getting-started/

Most of the articles about debugging advise you to use `node-inspector`
but that seems like it no longer works as of node 8.

I tried to figure out a way to pass the argument directly to `node` and
add a command to `scripts` in package.json but I didn't succeed.