Change copy assignments to use bulk create
Without a bulk create too many operations were being created in a time
frame and it was erroring.

Change the logic to only copy assignments for one week period. There was
a bug there other way and the date math got very confusing and the
feature was hard to logic about. Because there aren't clear guidelines I
just simplified the feature to remove the bug.

Bulk creation docs:
https://airtable.com/developers/blocks/api/models/Table#createRecordsAsync