Introduce an editorconfig (#39)
The [editorconfig] file alerts editors to the desired indentation
conventions for the project. [Plugins] exist for many editors.

[editorconfig]: https://editorconfig.org/
[Plugins]: https://editorconfig.org/#download

The goal is to match what we have configured for ESlint: four space
indentation, using spaces not tabs. Additionally we mandate the typical
unix conventions: newline is \n, and the final line ends with a newline.